requirements.txt
streamlit
mysql-connector-python
pandas
streamlit_authenticator